Code snippet to help clean unstructured CPU Name
The raw data from Geekbench.com somehow looks confusing, especially the CPU names. A CPU might have many variants based on their clock speed and cores. But in the reality they are have standard core, and clock speed is a variation that should be able to be ignored. I create a simple function in this file by combining compile and sub methods in regex and apply method in  pandas in order to clean CPU name in my dataset and standardized my dataset.